add: Blacklist expandable row, add shadows
